Update mock locket fixtures and server
This removes the fixture files and replaces it with an exported
SetupLocketFixtures/Cleanup funcs that write/remove files to a temp directory.
This allows a broker using this library to create a mock locket client and
server. This also updates the mock locket server so it can be initialised
and run during broker tests.