Add better broker api client
The existing BrokerTester in this repo is not brilliant, because it
just returns raw `*httptest.ResponseRecorder`s for all operations,
leaving the caller to parse the body as JSON and pull out the fields
they're interested in.

I found a better client in paas-rds-broker here:
https://github.com/alphagov/paas-rds-broker/blob/master/ci/helpers/brokerapi_helper.go
and thought it should be moved to somewhere that more people could use
it.