Make context timeout configurable
This adds a ContextTimeoutSeocnds field to API to allow callers to
customize the context timeout.  The motivation is
https://github.com/alphagov/paas-sqs-broker/pull/18 – we need a longer
timeout when trying to create a binding using CloudFormation.

It's possible to override it by using an init() func to change the
value of broker.DefaultContextTimeout but this is hackery.  Much
better to make it directly configurable.